+++ tests/testthat/examples/fastcpd_ar.R
@@ -5,7 +5,7 @@ for (i in 1:600) {
   x[i + 3] <- 0.6 * x[i + 2] - 0.2 * x[i + 1] + 0.1 * x[i] + rnorm(1, 0, 3)
 }
 for (i in 601:1000) {
-  x[i + 1] <- 0.3 * x[i + 2] + 0.4 * x[i + 1] + 0.2 * x[i] + rnorm(1, 0, 3)
+  x[i + 3] <- 0.3 * x[i + 2] + 0.4 * x[i + 1] + 0.2 * x[i] + rnorm(1, 0, 3)
 }
 result <- fastcpd.ar(x[3 + seq_len(n)], 3)
 summary(result)